Extracted the island logic so that it can be applied independently of the underlying implementations of Island/Hexalot/Spot
