+++ client/src/galapagotchi.tsx
@@ -15,7 +15,7 @@ import { createFabricKernel, FabricKernel } from "./body/fabric-kernel"
 import { Physics } from "./body/physics"
 import { INITIAL_JOINT_COUNT, MAX_POPULATION } from "./gotchi/evolution"
 import { Island } from "./island/island"
-import { Surface } from "./island/spot"
+import { Surface } from "./island/island-logic"
 import { IAppState, logString } from "./state/app-state"
 import { IStorage } from "./storage/storage"
 import { ActionsPanel } from "./view/actions-panel"
