+++ client/src/island/island.ts
@@ -11,19 +11,12 @@ import { IAppState, Mode } from "../state/app-state"
 import { Transition } from "../state/transition"
 import { IStorage } from "../storage/storage"
 
-import { Hexalot, hexalotTreeString } from "./hexalot"
-import { ADJACENT, BRANCH_STEP, HEXALOT_SHAPE, STOP_STEP } from "./shapes"
-import { coordSort, equals, ICoords, plus, Spot, spotsToString, Surface, zero } from "./spot"
+import { Hexalot } from "./hexalot"
+import { constructIsland, equals, ICoords, IIsland, IslandData, plus } from "./island-logic"
+import { ADJACENT, HEXALOT_SHAPE } from "./shapes"
+import { Spot } from "./spot"
 
-export interface IslandData {
-    name: string
-    hexalots: string
-    spots: string
-}
-
-const sortSpotsOnCoord = (a: Spot, b: Spot): number => coordSort(a.coords, b.coords)
-
-export class Island {
+export class Island implements IIsland {
     public readonly name: string
     public spots: Spot[] = []
     public hexalots: Hexalot[] = []
@@ -31,7 +24,7 @@ export class Island {
     public vacantHexalot?: Hexalot
 
     constructor(islandData: IslandData, readonly gotchiFactory: IGotchiFactory, storage: IStorage, subject: Subject<IAppState>, nonce: number) {
-        this.apply(islandData)
+        constructIsland(islandData, this)
         this.name = islandData.name
         const island = this
         const mode = Mode.Visiting
@@ -80,74 +73,7 @@ export class Island {
             .multiplyScalar(1 / this.spots.length)
     }
 
-    public get data(): IslandData {
-        if (!this.islandIsLegal) {
-            throw new Error("Saving illegal island")
-        }
-        this.spots.sort(sortSpotsOnCoord)
-        return {
-            name: this.name,
-            hexalots: hexalotTreeString(this.hexalots),
-            spots: spotsToString(this.spots),
-        } as IslandData
-    }
-
-    // ================================================================================================
-
-    private apply(data: IslandData): void {
-        let hexalot: Hexalot | undefined = this.getOrCreateHexalot(undefined, zero)
-        const stepStack = data.hexalots.split("").reverse().map(stepChar => Number(stepChar))
-        const hexalotStack: Hexalot[] = []
-        while (stepStack.length > 0) {
-            const step = stepStack.pop()
-            switch (step) {
-                case STOP_STEP:
-                    hexalot = hexalotStack.pop()
-                    break
-                case BRANCH_STEP:
-                    if (hexalot) {
-                        hexalotStack.push(hexalot)
-                    }
-                    break
-                case 1:
-                case 2:
-                case 3:
-                case 4:
-                case 5:
-                case 6:
-                    if (hexalot) {
-                        hexalot = this.hexalotAroundSpot(hexalot.spots[step])
-                    }
-                    break
-                default:
-                    console.error("Error step")
-            }
-        }
-        const hexChars = data.spots ? data.spots.split("") : []
-        const numbers = hexChars.map(hexChar => parseInt(hexChar, 16))
-        const booleanArrays = numbers.map(nyb => {
-            const b0 = (nyb & 8) !== 0
-            const b1 = (nyb & 4) !== 0
-            const b2 = (nyb & 2) !== 0
-            const b3 = (nyb & 1) !== 0
-            return [b0, b1, b2, b3]
-        })
-        const landStack = [].concat.apply([], booleanArrays).reverse()
-        this.spots.sort(sortSpotsOnCoord)
-        if (landStack.length) {
-            this.spots.forEach(spot => {
-                const land = landStack.pop()
-                spot.surface = land ? Surface.Land : Surface.Water
-            })
-        } else if (this.hexalots.length === 1) {
-            const singleHexalot = this.hexalots[0]
-            singleHexalot.spots.map(spot => spot.surface = Math.random() > 0.5 ? Surface.Land : Surface.Water)
-            singleHexalot.spots[0].surface = Surface.Land
-        }
-        this.hexalots.forEach(lot => lot.refreshId())
-    }
-
-    private hexalotAroundSpot(spot: Spot): Hexalot {
+    public hexalotAroundSpot(spot: Spot): Hexalot {
         const hexalotParent = spot.adjacentHexalots.reduce(
             (parent: Hexalot | undefined, candiate: Hexalot) => {
                 if (parent && parent.nonce >= candiate.nonce) {
@@ -160,7 +86,7 @@ export class Island {
         return this.getOrCreateHexalot(hexalotParent, spot.coords)
     }
 
-    private getOrCreateHexalot(parent: Hexalot | undefined, coords: ICoords): Hexalot {
+    public getOrCreateHexalot(parent: Hexalot | undefined, coords: ICoords): Hexalot {
         const existing = this.hexalots.find(existingHexalot => equals(existingHexalot.coords, coords))
         if (existing) {
             return existing
@@ -171,6 +97,8 @@ export class Island {
         this.hexalots.push(hexalot)
         return hexalot
     }
+
+    // ================================================================================================
 
     private getOrCreateSpot(coords: ICoords): Spot {
         const existing = this.getSpot(coords)
