+++ client/src/storage/browser-storage.ts
@@ -5,7 +5,8 @@
 
 import { fromOptionalGenomeData, IGenomeData } from "../genetics/genome"
 import { Hexalot } from "../island/hexalot"
-import { Island, IslandData } from "../island/island"
+import { Island } from "../island/island"
+import { extractIslandData, IslandData } from "../island/island-logic"
 import { IJourneyData } from "../island/journey"
 
 import { IStorage } from "./storage"
@@ -34,9 +35,9 @@ export class BrowserStorage implements IStorage {
     public async claimHexalot(island: Island, hexalot: Hexalot, genomeData: IGenomeData): Promise<IslandData | undefined> {
         hexalot.genome = fromOptionalGenomeData(genomeData)
         this.setGenomeData(hexalot, genomeData)
-        const islandData = island.data
+        const islandData = extractIslandData(island)
         this.storage.setItem(islandData.name, JSON.stringify(islandData))
-        return Promise.resolve(island.data)
+        return Promise.resolve(islandData)
     }
 
     public async getGenomeData(hexalot: Hexalot): Promise<IGenomeData | undefined> {
