+++ client/src/storage/storage.ts
@@ -5,7 +5,8 @@
 
 import { IGenomeData } from "../genetics/genome"
 import { Hexalot } from "../island/hexalot"
-import { Island, IslandData } from "../island/island"
+import { Island } from "../island/island"
+import { IslandData } from "../island/island-logic"
 import { IJourneyData } from "../island/journey"
 
 export interface IStorage {
